move some of the core pieces around so that it's easier to follow
