+++ src/server/config.ts
@@ -1,0 +1,7 @@
+export interface Config {
+  basePath: string;
+  openApiPath: string;
+  port: number;
+  proxyEnabled: boolean;
+  proxyUrl: string;
+}
