--- test/server/core.test.js
@@ -1,150 +1,0 @@
-import http from "node:http";
-
-import Koa from "koa";
-import supertest from "supertest";
-
-import { core } from "../../src/server/core.ts";
-import { withTemporaryFiles } from "../lib/with-temporary-files.ts";
-
-describe("integration test", () => {
-  it("finds a path", async () => {
-    const app = new Koa();
-    const request = supertest(app.callback());
-    const files = {
-      "paths/hello.mjs": `
-        export async function GET() {
-          return await Promise.resolve({ body: "GET /hello" });
-        }
-      `,
-      "paths/hello/world.mjs": `
-        export async function POST() {
-          return await Promise.resolve({ body: "POST /hello/world" });
-        }
-      `,
-      "paths/teapot.mjs": `
-      export async function POST() {
-        return await Promise.resolve({ status: 418, body: "I am a teapot." });
-      }
-    `,
-    };
-
-    await withTemporaryFiles(files, async (basePath) => {
-      const { koaMiddleware, start, stop } = await core(basePath);
-
-      app.use(koaMiddleware);
-
-      await start();
-
-      const getResponse = await request.get("/hello");
-      const postResponse = await request.post("/hello/world");
-      const teapotResponse = await request.post("/teapot");
-
-      expect(getResponse.text).toBe("GET /hello");
-      expect(postResponse.text).toBe("POST /hello/world");
-      expect(teapotResponse.statusCode).toBe(418);
-
-      await stop();
-    });
-  });
-
-  it("uses an OpenAPI document to generate a random response", async () => {
-    const app = new Koa();
-    const request = supertest(app.callback());
-    const files = {
-      "openapi.yaml": `
-        openapi: 3.0.0
-        info:
-          title: Counterfact
-          version: 1.0.0
-        paths:
-          /hello:
-            get:
-              responses:
-                200:
-                  content:
-                    text/plain:
-                      schema:
-                        $ref: "#/components/schemas/Hello"
-        components:
-          schemas:
-            Hello: 
-              type: string
-              examples: 
-                - "hello"
-      `,
-
-      "paths/hello.mjs": `
-        export async function GET({response}) {
-          return response[200].random();
-        }
-      `,
-    };
-
-    await withTemporaryFiles(files, async (basePath) => {
-      const { koaMiddleware, start, stop } = await core(
-        basePath,
-        `${basePath}/openapi.yaml`,
-      );
-
-      await start();
-
-      app.use(koaMiddleware);
-
-      const getResponse = await request.get("/hello");
-
-      expect(getResponse.text).toBe("hello");
-
-      await stop();
-    });
-  });
-
-  it("proxies when a proxyUrl is present and the proxy is enabled", async () => {
-    const proxyTarget = http.createServer((req, res) => {
-      res.writeHead(200, {
-        "Content-Type": "text/plain",
-      });
-
-      res.end("I am proxy!\n");
-    });
-
-    proxyTarget.listen(0);
-
-    const app = new Koa();
-    const request = supertest(app.callback());
-    const files = {
-      "paths/hello.mjs": `
-        export async function GET() {
-          return await Promise.resolve({ body: "GET /hello" });
-        }
-      `,
-    };
-
-    await withTemporaryFiles(files, async (basePath) => {
-      const { koaMiddleware, start, stop } = await core(
-        basePath,
-        `${basePath}/openapi.yaml`,
-        {
-          proxyEnabled: true,
-          proxyUrl: `http://localhost:${proxyTarget.address().port}`,
-        },
-      );
-
-      app.use(koaMiddleware);
-
-      await start();
-
-      const getResponse = await request.get("/hello");
-
-      expect(getResponse.text).toBe("I am proxy!\n");
-
-      await stop();
-
-      await proxyTarget.unref();
-    });
-
-    // eslint-disable-next-line promise/avoid-new
-    await new Promise((resolve) => {
-      proxyTarget.close(resolve);
-    });
-  });
-});
